refactor: entity와 enum 패키지 분리
