+++ src/main/java/com/umc/halo/domain/member/enums/Provider.java
@@ -1,4 +1,4 @@
-package com.umc.halo.domain.member.entity;
+package com.umc.halo.domain.member.enums;
 
 public enum Provider {
     GOOGLE, // 구글 로그인
