+++ src/main/java/com/umc/halo/domain/record/enums/Status.java
@@ -1,4 +1,4 @@
-package com.umc.halo.domain.record.entity;
+package com.umc.halo.domain.record.enums;
 
 public enum Status {
     DRAFT,     // 임시저장
